QATlib: 23.11.0 release
Changes from 23.08.0 to 23.11.0:
* Support DC NS (NoSession) APIs
* Support Symmetric Crypto SM3 & SM4
* Support Asymmetric Crypto SM2
* DC compressBound APIs
* Bug Fixes. See Resolved section in README.md

+++ quickassist/lookaside/access_layer/src/common/utils/lac_buffer_desc.c
@@ -290,9 +290,10 @@ CpaStatus LacBuffDesc_FlatBufferVerify(
         if (!LAC_ADDRESS_ALIGNED(pUserFlatBuffer->pData,
                                  alignmentShiftExpected))
         {
-            LAC_INVALID_PARAM_LOG1("FlatBuffer not aligned as expected."
-                                   " Expected alignment %u Bytes.",
-                                   (uintptr_t)(1 << alignmentShiftExpected));
+            LAC_INVALID_PARAM_LOG1(
+                "FlatBuffer not aligned as expected."
+                " Expected alignment %lu Bytes.",
+                (unsigned long)(1 << alignmentShiftExpected));
             return CPA_STATUS_INVALID_PARAM;
         }
     }
@@ -323,9 +324,10 @@ CpaStatus LacBuffDesc_FlatBufferVerifyNull(
         if (!LAC_ADDRESS_ALIGNED(pUserFlatBuffer->pData,
                                  alignmentShiftExpected))
         {
-            LAC_INVALID_PARAM_LOG1("FlatBuffer not aligned as expected."
-                                   " Expected alignment %u Bytes.",
-                                   (uintptr_t)(1 << alignmentShiftExpected));
+            LAC_INVALID_PARAM_LOG1(
+                "FlatBuffer not aligned as expected."
+                " Expected alignment %lu Bytes.",
+                (unsigned long)(1 << alignmentShiftExpected));
             return CPA_STATUS_INVALID_PARAM;
         }
     }
